+++ src/MultiCascader/Dropdown.js
@@ -7,14 +7,17 @@ import { shallowEqualArray } from 'rsuite-utils/lib/utils';
 import { polyfill } from 'react-lifecycles-compat';
 
 import DropdownMenu from './DropdownMenu';
+import Checkbox from '../Checkbox';
 import createUtils from './utils';
 import type { Placement } from '../utils/TypeDefinition';
 
 import { defaultProps, prefix, getUnhandledProps, createChainedFunction } from '../utils';
+import { flattenTree, getNodeParents } from '../utils/treeUtils';
 
 import {
   PickerToggle,
   MenuWrapper,
+  SearchBar,
   SelectedElement,
   PickerToggleTrigger,
   getToggleWrapperClassName,
@@ -45,7 +48,9 @@ type Props = {
   value?: any[],
   defaultValue?: any[],
   placeholder?: string,
+  onSearch?: (searchKeyword: string, event: DefaultEvent) => void,
   onChange?: (value: any, event: DefaultEvent) => void,
+  onClean?: (event: DefaultEvent) => void,
   onOpen?: () => void,
   onClose?: () => void,
   onHide?: () => void,
@@ -75,7 +80,8 @@ type Props = {
   menuWidth?: number,
   menuHeight?: number,
   style?: Object,
-  uncheckableItemValues?: any[]
+  uncheckableItemValues?: any[],
+  searchable?: boolean
 };
 
 type State = {
@@ -86,7 +92,8 @@ type State = {
   items?: any[],
   data: any[],
   flattenData: any[],
-  active?: boolean
+  active?: boolean,
+  searchKeyword: string
 };
 
 class Dropdown extends React.Component<Props, State> {
@@ -101,15 +108,17 @@ class Dropdown extends React.Component<Props, State> {
     labelKey: 'label',
     locale: {
       placeholder: 'Select',
-      checkAll: 'All'
+      checkAll: 'All',
+      searchPlaceholder: 'Search',
+      noResultsText: 'No results found'
     },
     cleanable: true,
+    searchable: true,
     countable: true,
     placement: 'bottomLeft'
   };
 
   static utils = {};
-
   isControlled = null;
   constructor(props: Props) {
     super(props);
@@ -117,6 +126,7 @@ class Dropdown extends React.Component<Props, State> {
     const { data, value, defaultValue } = props;
     const initState = {
       data,
+      searchKeyword: '',
       prevValue: value,
       value: defaultValue,
       selectNode: null,
@@ -127,7 +137,7 @@ class Dropdown extends React.Component<Props, State> {
     };
 
     Dropdown.utils = createUtils(props);
-    const flattenData = Dropdown.utils.flattenNodes(data);
+    const flattenData = flattenTree(data, props.childrenKey);
 
     this.isControlled = !_.isUndefined(value);
     this.state = {
@@ -170,7 +180,7 @@ class Dropdown extends React.Component<Props, State> {
 
     if (isChangedData || isChangedValue) {
       if (isChangedData) {
-        flattenData = Dropdown.utils.flattenNodes(data);
+        flattenData = flattenTree(data, nextProps.childrenKey);
       }
 
       /**
@@ -236,6 +246,14 @@ class Dropdown extends React.Component<Props, State> {
     onChange && onChange(value, event);
   };
 
+  handleChangeForSearchItem = (
+    value: any,
+    checked: boolean,
+    event: SyntheticInputEvent<HTMLInputElement>
+  ) => {
+    this.handleCheck(value, event, checked);
+  };
+
   handleSelect = (node: Object, cascadeItems, activePaths: any[], event: DefaultEvent) => {
     const { onSelect, valueKey } = this.props;
 
@@ -247,6 +265,14 @@ class Dropdown extends React.Component<Props, State> {
 
     onSelect &&
       onSelect(node, activePaths, createConcatChildrenFunction(node, node[valueKey]), event);
+  };
+
+  handleSearch = (searchKeyword: string, event: DefaultEvent) => {
+    const { onSearch } = this.props;
+    this.setState({
+      searchKeyword
+    });
+    onSearch && onSearch(searchKeyword, event);
   };
 
   trigger = null;
@@ -301,25 +327,125 @@ class Dropdown extends React.Component<Props, State> {
     const { onClose } = this.props;
     onClose && onClose();
     this.setState({
+      searchKeyword: '',
       active: false
     });
   };
 
   addPrefix = (name: string) => prefix(this.props.classPrefix)(name);
 
+  getSearchResult() {
+    const { labelKey, valueKey, uncheckableItemValues = [] } = this.props;
+    const { searchKeyword, flattenData } = this.state;
+
+    const items = [];
+    const result = flattenData.filter(item => {
+      if (uncheckableItemValues.some(value => item[valueKey] === value)) {
+        return false;
+      }
+
+      if (item[labelKey].match(new RegExp(searchKeyword, 'i'))) {
+        return true;
+      }
+      return false;
+    });
+
+    for (let i = 0; i < result.length; i++) {
+      items.push(result[i]);
+
+      if (i === 99) {
+        return items;
+      }
+    }
+    return items;
+  }
+
+  renderSearchRow = (item: Object, key: number) => {
+    const { labelKey, valueKey, cascade, disabledItemValues = [] } = this.props;
+    const { searchKeyword } = this.state;
+    const values = this.getValue();
+    const nodes = getNodeParents(item);
+    const regx = new RegExp(searchKeyword, 'ig');
+    const labelElements = [];
+
+    let a = item[labelKey].split(regx);
+    let b = item[labelKey].match(regx);
+
+    for (let i = 0; i < a.length; i++) {
+      labelElements.push(a[i]);
+      if (b[i]) {
+        labelElements.push(<strong key={i}>{b[i]}</strong>);
+      }
+    }
+
+    nodes.push({
+      ...item,
+      [labelKey]: labelElements
+    });
+
+    const active = values.some(value => nodes.some(node => node[valueKey] === value));
+    const disabled = disabledItemValues.some(value => nodes.some(node => node[valueKey] === value));
+    const itemClasses = classNames(this.addPrefix('cascader-row'), {
+      [this.addPrefix('cascader-row-disabled')]: disabled
+    });
+
+    return (
+      <div key={key} className={itemClasses}>
+        <Checkbox
+          disabled={disabled}
+          checked={active}
+          value={item}
+          indeterminate={cascade && !active && Dropdown.utils.isSomeChildChecked(item, values)}
+          onChange={this.handleChangeForSearchItem}
+        >
+          <span className={this.addPrefix('cascader-cols')}>
+            {nodes.map((node, index) => (
+              <span key={`col-${index}`} className={this.addPrefix('cascader-col')}>
+                {node[labelKey]}
+              </span>
+            ))}
+          </span>
+        </Checkbox>
+      </div>
+    );
+  };
+
+  renderSearchResultPanel() {
+    const { locale } = this.props;
+    const { searchKeyword } = this.state;
+
+    if (searchKeyword === '') {
+      return null;
+    }
+
+    const items = this.getSearchResult();
+    return (
+      <div className={this.addPrefix('cascader-search-panel')}>
+        {items.length ? (
+          items.map(this.renderSearchRow)
+        ) : (
+          <div className={this.addPrefix('none')}>{locale.noResultsText}</div>
+        )}
+      </div>
+    );
+  }
+
   renderDropdownMenu() {
-    const { items, activePaths } = this.state;
+    const { items, activePaths, searchKeyword } = this.state;
     const {
       renderMenu,
       placement,
       renderExtraFooter,
       menuClassName,
       menuStyle,
-      classPrefix
+      classPrefix,
+      searchable,
+      locale
     } = this.props;
 
     const classes = classNames(
       this.addPrefix('cascader-menu'),
+      this.addPrefix('multi-cascader-menu'),
       this.addPrefix(`placement-${_.kebabCase(placement)}`),
       menuClassName
     );
@@ -328,17 +454,30 @@ class Dropdown extends React.Component<Props, State> {
 
     return (
       <MenuWrapper className={classes} style={menuStyle}>
-        <DropdownMenu
-          {...menuProps}
-          classPrefix={classPrefix}
-          ref={this.bindMenuContainerRef}
-          cascadeItems={items}
-          cascadePathItems={activePaths}
-          value={this.getValue()}
-          onSelect={this.handleSelect}
-          onCheck={this.handleCheck}
-          renderMenu={renderMenu}
-        />
+        {searchable && (
+          <SearchBar
+            placeholder={locale.searchPlaceholder}
+            onChange={this.handleSearch}
+            value={searchKeyword}
+          />
+        )}
+
+        {this.renderSearchResultPanel()}
+
+        {searchKeyword === '' && (
+          <DropdownMenu
+            {...menuProps}
+            classPrefix={classPrefix}
+            ref={this.bindMenuContainerRef}
+            cascadeItems={items}
+            cascadePathItems={activePaths}
+            value={this.getValue()}
+            onSelect={this.handleSelect}
+            onCheck={this.handleCheck}
+            renderMenu={renderMenu}
+          />
+        )}
+
         {renderExtraFooter && renderExtraFooter()}
       </MenuWrapper>
     );
@@ -358,6 +497,7 @@ class Dropdown extends React.Component<Props, State> {
       style,
       onEnter,
       onExited,
+      onClean,
       countable,
       cascade,
       ...rest
@@ -405,7 +545,7 @@ class Dropdown extends React.Component<Props, State> {
           <PickerToggle
             {...unhandled}
             componentClass={toggleComponentClass}
-            onClean={this.handleClean}
+            onClean={createChainedFunction(this.handleClean, onClean)}
             cleanable={cleanable && !disabled}
             hasValue={hasValue}
             active={this.state.active}