+++ src/SelectPicker/Dropdown.js
@@ -62,15 +62,16 @@ type Props = {
   onSelect?: (value: any, item: Object, event: DefaultEvent) => void,
   onGroupTitleClick?: DefaultEventFunction,
   onSearch?: (searchKeyword: string, event: DefaultEvent) => void,
+  onClean?: (event: DefaultEvent) => void,
   onOpen?: () => void,
   onClose?: () => void,
   onHide?: () => void,
-  onEnter?: Function,
-  onEntering?: Function,
-  onEntered?: Function,
-  onExit?: Function,
-  onExiting?: Function,
-  onExited?: Function,
+  onEnter?: () => void,
+  onEntering?: () => void,
+  onEntered?: () => void,
+  onExit?: () => void,
+  onExiting?: () => void,
+  onExited?: () => void,
   /**
    * group by key in `data`
    */
@@ -479,6 +480,7 @@ class Dropdown extends React.Component<Props, State> {
       style,
       onEntered,
       onExited,
+      onClean,
       ...rest
     } = this.props;
 
@@ -514,7 +516,7 @@ class Dropdown extends React.Component<Props, State> {
           <PickerToggle
             {...unhandled}
             ref={this.bindToggleRef}
-            onClean={this.handleClean}
+            onClean={createChainedFunction(this.handleClean, onClean)}
             onKeyDown={this.handleKeyDown}
             componentClass={toggleComponentClass}
             cleanable={cleanable && !disabled}