+++ src/Table.js
@@ -6,10 +6,9 @@ import { Table, Column, Cell, HeaderCell } from 'rsuite-table';
 import withLocale from './IntlProvider/withLocale';
 import TablePagination from './TablePagination';
 
-const EnhancedLocaleTable = compose(
-  withLocale(['Table']),
-  defaultProps({ loadAnimation: true })
-)(Table);
+const EnhancedLocaleTable = compose(withLocale(['Table']), defaultProps({ loadAnimation: true }))(
+  Table
+);
 
 setStatic('Column', Column)(EnhancedLocaleTable);
 setStatic('Cell', Cell)(EnhancedLocaleTable);