+++ types/MultiCascader.d.ts
@@ -32,6 +32,15 @@ export interface MultiCascaderProps extends FormControlPickerProps<any[]> {
 
   /** Called when the option is selected */
   onSelect?: (value: any, activePaths: any[], event: React.SyntheticEvent<HTMLElement>) => void;
+
+  /** Called when clean */
+  onClean?: (event: React.SyntheticEvent<HTMLElement>) => void;
+
+  /** Called when searching */
+  onSearch?: (searchKeyword: string, event: React.SyntheticEvent<HTMLElement>) => void;
+
+  /** Whether dispaly search input box */
+  searchable?: boolean;
 }
 
 declare const MultiCascader: React.ComponentType<MultiCascaderProps>;